requirements.txt
# Core deep learning
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Topological data analysis
gudhi>=3.8.0
persim>=0.3.1
ripser>=0.6.4
scikit-tda>=1.0.0

# Hyperbolic geometry
geoopt>=0.5.0

# Formal verification
z3-solver>=4.12.0

# Multi-objective optimization
pymoo>=0.6.0
botorch>=0.9.0  # For EHVI

# Monitoring & observability
prometheus-client>=0.17.0
grafana-api>=1.0.3

# Data processing
librosa>=0.10.0  # Audio processing
opencv-python>=4.8.0  # Video processing
timm>=0.9.0  # Vision transformers
transformers>=4.30.0  # Text tokenization

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0  # Optional experiment tracking
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
